# perf: MakeDefaultConfig once