@chlunde 2023-03-12
rebased for Go 1.20/CI fix