@shapirus 2023-11-03
This change has also significantly improved performance in Component-specific tests (using multiple invocations of kustomize binary), which is very nice:

```
Starting kustomize benchmark on Linux x86_64
kustomize versions: 
  5b51722
  2fda12d
iterations per test: 1000
tests: 
  1_no-patches
  2_patches-json6902
  3_patches-strategic-merge
  4_no-patches-unknown-kind
  5_component-no-base-no-patches
  6_component-json6902-over-base
  7_component-PSM-over-base
time unit: seconds

             test: 1   test: 2   test: 3   test: 4   test: 5   test: 6   test: 7
   5b51722      5.61      5.85     58.91      5.23     10.14     14.63     66.61
   2fda12d      5.47      5.73     58.90      5.30      5.57      8.08     60.52

```

Now only if we could improve patchesStrategicMerge performance... :)
Apparently it's the only thing left that's still using expensive openapi calls (https://github.com/kubernetes-sigs/kustomize/issues/5422#issuecomment-1783527597).